README.md
This project is a website called "Jimma Coffee Express." The website is built using HTML, CSS, and PHP. It allows users to purchase coffee from Jimma and have it delivered to various cities across Ethiopia. 

This website offers an easy and convenient way to enjoy high-quality Jimma coffee from the comfort of your home.